pyproject.toml
[project]
name = "quantum-phoenix"
version = "1.0.0"
description = "Phoenix-instrumented quantum physics RAG demo with agentic workflows"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.3.0",
    "chromadb>=0.4.18",
    "sentence-transformers>=2.7.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "arize-phoenix>=4.0.0",
    "arize-phoenix-otel>=0.0.8",
    "openinference-instrumentation>=0.1.0",
    "openinference-semantic-conventions>=0.1.0",
    "opentelemetry-api>=1.21.0",
    "opentelemetry-sdk>=1.21.0",
    "beautifulsoup4>=4.12.2",
    "requests>=2.31.0",
    "urllib3>=2.1.0",
    "lxml>=5.0.0",
    "sqlalchemy>=2.0.36",
    "aiosqlite>=0.19.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "httpx>=0.25.2",
    "tenacity>=8.2.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "scripts"]

[dependency-groups]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "black>=23.11.0",
    "ruff>=0.1.6",
]
